Redefine ioctl(FS_IOC_[SG]ETFLAGS) to take an int*
/usr/include/linux/fs.h: #define FS_IOC_SETFLAGS _IOW('f', 2, long)

Apparently this is known issue [1], but a fix does not seem to be
in sight. Let's fix it on our end, so that valgrind does not throw
false positives:

==31046== Syscall param ioctl(generic) points to uninitialised byte(s)
==31046==    at 0x5BB18C7: ioctl (syscall-template.S:84)
==31046==    by 0x41B791: ca_decoder_finalize_child (cadecoder.c:3656)
==31046==    by 0x41C0BE: ca_decoder_step_node (cadecoder.c:3875)
==31046==    by 0x41C9ED: ca_decoder_step (cadecoder.c:4109)
==31046==    by 0x408496: ca_sync_step_decode (casync.c:1988)
==31046==    by 0x4094F6: ca_sync_step (casync.c:2482)
==31046==    by 0x40366B: main (test-casync.c:75)
==31046==  Address 0xffefff5fc is on thread 1's stack
==31046==  in frame #1, created by ca_decoder_finalize_child (cadecoder.c:3252)

While at it, let's change the internal type from unsigned to int, since that's
what the kernel really uses.

[1] http://linux-fsdevel.vger.kernel.narkive.com/eUZdzNjw/argument-type-for-fs-ioc-getflags-fs-ioc-setflags-ioctls

+++ src/util.c
@@ -672,10 +672,10 @@ char* ls_format_mode(mode_t m, char ret[LS_FORMAT_MODE_MAX]) {
         return ret;
 }
 
-char *ls_format_chattr(unsigned flags, char ret[LS_FORMAT_CHATTR_MAX]) {
+char *ls_format_chattr(int flags, char ret[LS_FORMAT_CHATTR_MAX]) {
 
         static const struct {
-                unsigned flag;
+                int flag;
                 char code;
         } table[] = {
                 { FS_SYNC_FL,        'S' },
@@ -692,7 +692,7 @@ char *ls_format_chattr(unsigned flags, char ret[LS_FORMAT_CHATTR_MAX]) {
 
         size_t i;
 
-        if (flags == (unsigned) -1)
+        if (flags == -1)
                 return NULL;
 
         assert(ELEMENTSOF(table) == LS_FORMAT_CHATTR_MAX-1);
